back_end/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.3
coveralls==1.11.1
defusedxml==0.6.0
Django==3.0.3
django-debug-toolbar==2.2
django-widget-tweaks==1.4.5
docopt==0.6.2
Pillow==7.0.0
PyJWT==1.7.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
selenium==3.141.0
six==1.14.0
sqlparse==0.3.0
bleach==3.1.0
faker==4.0.0
django-cors-headers==3.2.1
djangorestframework==3.11.0
markdown==3.2.1
django-filter==2.2.0
django-rest-knox==4.1.0
